Register Page
Finished the Register Page. You can create an account, and it will automatically tell you if you're password is too short, if it needs a number, if it needs a letter minimum, and stuff like that. It checks if the password confirmation matches the password that you chose. Still needs a validation for password characters, IT SHOULD NOT contain spaces. Will update that next.