Local Register Backend
Corrected valid password conditions (should not contain spaces).
Started working on backend for Flask server, created api for login and register, created data base fallnik.db for user accounts, passwords are encrypted, with every new user we generate a random salt that we concatenate to the password, and that result is an input for a sha256 hash function. Also created real time check functions for the register page, it alerts with a message that an username or email already exists before you try to submit your registration.