FallnikAI Chatbot, AI Summary for .txt, .pdf, .docx files, hidden important keys
Developed FallnikAI Page, still in a early stage, but every functionality works. Added FallnikAI bot, an AI that can answer questions related to different topics helping students learn easier. It keeps in mind the context of the lesson that you are summarizing in the Upload and Summarize function, and it can answer questions related to the topic and much more. Upload and Summarize works, you can upload any text file, pdf or docx that lets say contains a lesson, and it will summarize the content of the file and give you a clear and easy to learn summary. It works fine for now. Still have to work on the design of the page. I've also hidden OpenAI key and  the mail generated key from my personal account so they will not be able to be seen in this public repository.