Completed Spotify Account Connection
Added a mini player, the page dedicated to spotify works fine.
I had to generate a new openai key, the old one was invalidated so the ai was not working anymore, i organized the code a bit, created a MusicPlayerContext, from which i take the functions that i need for other components and also for the main MusicPage. Music now works across pages.